Migrate query() function.
